Kill handles no-input case.
git-svn-id: https://hotwire-shell.googlecode.com/svn/trunk@1066 79f8d753-db37-0410-980a-393e3d9e3e3e

+++ hotwire/builtins/kill.py
@@ -96,8 +96,9 @@ class KillBuiltin(Builtin):
                     break
         for arg in imap(int, args):
             os.kill(arg, signum)
-        for arg in context.input:
-            os.kill(arg.pid, signum)
+        if context.input is not None:
+            for arg in context.input:
+                os.kill(arg.pid, signum)
         return []
         
 BuiltinRegistry.getInstance().register_hotwire(KillBuiltin())
